EOS-27315: hctl bootstrap is failing in latest hare main
hctl bootstrap hangs in main,
```
[root@ssc-vm-g2-rhev4-1947 ~]# hctl bootstrap --mkfs --xprt libfab ~/1n.yaml
2022-01-07 10:44:17: Generating cluster configuration... OK
2022-01-07 10:44:20: Starting Consul server on this node......... OK
2022-01-07 10:44:29: Importing configuration into the KV store... OK
2022-01-07 10:44:31: Starting Consul on other nodes...Consul ready on all nodes
2022-01-07 10:44:31: Updating Consul configuraton from the KV store... OK
2022-01-07 10:44:33: Waiting for the RC Leader to get elected.... OK
2022-01-07 10:44:36: Starting Motr (phase1, mkfs)...
```

Because config_path kv is missing is missing in Consul, config_path is added through hare mini provisioner, which is not invoked as part of hctl bootstrap.
```
Jan 07 10:49:18 ssc-vm-g2-rhev4-1947.colo.seagate.com hare-hax[6027]: 2022-01-07 10:49:18,919 [DEBUG] {MainThread} Got HAConsistencyException: config path not present in consul while invokingJan 07 10:49:23 ssc-vm-g2-rhev4-1947.colo.seagate.com hare-hax[6027]: 2022-01-07 10:49:23,924 [DEBUG] {MainThread} KVGET key=config_path, kwargs={}
```

This is a regression from commit 28869a4f.

Solution:
- Move `MessageBus.init` to ha module, thus the relevant code will be
invoked only if deployed with cortx-ha component.
- Install m0ping utility as part of hare `make devinstall`
